+++ langs/tail.datetime-tr.js
@@ -1,3 +1,17 @@
+/*
+ |  tail.datetime - A vanilla JavaScript DateTime Picker without dependencies!
+ |  @file       ./langs/tail.datetime-tr.js
+ |  @author     SamBrishes <sam@pytes.net>
+ |  @version    0.4.10 - Beta
+ |
+ |  @website    https://github.com/pytesNET/tail.DateTime
+ |  @license    X11 / MIT License
+ |  @copyright  Copyright © 2018 - 2019 SamBrishes, pytesNET <info@pytes.net>
+ */
+/*
+ |  Translator:     Murat Pala - (https://github.com/Prozexis)
+ |  GitHub:         https://github.com/pytesNET/tail.DateTime/pull/30
+ */
 ;(function(factory){
    if(typeof(define) == "function" && define.amd){
        define(function(){